refactor: dedupe renderer shims via shared crates (ecs/client/server)
- Add ecs_core, client_core, server_core crates
- Switch render_wgpu to re-export these crates; remove copied modules
- Move root src/{ecs,client,server} to crate re-exports
- Provide renderer-only collision via extension trait (server_ext)
- Build/clippy/tests all green
