+++ crates/render_wgpu/Cargo.toml
@@ -15,6 +15,9 @@ winit = "0.30.12"
 ra-assets = { version = "0.1.0", path = "../../shared/assets" }
 data_runtime = { version = "0.1.0", path = "../data_runtime" }
 ux_hud = { version = "0.1.0", path = "../ux_hud" }
+ecs_core = { version = "0.1.0", path = "../ecs_core" }
+client_core = { version = "0.1.0", path = "../client_core" }
+server_core = { version = "0.1.0", path = "../server_core" }
 rand = "0.9.2"
 rand_chacha = "0.9.0"
 serde = { version = "1.0.228", features = ["derive"] }
