--- crates/render_wgpu/src/client/mod.rs
@@ -1,3 +1,0 @@
-pub mod input;
-pub mod controller;
-
