+++ crates/server_core/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "server_core"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+glam = "0.30"
+
